- Domain models were modified. - The local database is restructured in order to store the necessary information. - The use cases layer was created to communicate with the ViewModels. - The message utility is added. - The dependencies module is created for use cases. - The following use cases are created: store search, list recent searches and list products. - The utility that allows uploading images to a Textview is created. - The product adapter is created. - The dependency manager is modified. - The main activity is related to the search for a product.

+++ app/src/main/java/com/jan/melichallenge/data/api/response/Attribute.kt
@@ -13,7 +13,7 @@ data class Attribute(
     @SerializedName("name")
     val name: String,
     @SerializedName("source")
-    val source: Int,
+    val source: Long,
     @SerializedName("value_id")
     val valueId: String?,
     @SerializedName("value_name")
